output kernel_src_path
Summary:
The changes made in the code include adding a new field to the output of the tool, which is the path to the source code of the kernel.

This field is added to the output of the tool to make it easier for developers to locate the source code of the kernel.

`kernel_src_path` will be used in custom module loading hack.

Reviewed By: FindHao

Differential Revision: D83608135

fbshipit-source-id: 4790ebe6df87de8823d6db929b0cbaf0c64d8fba